ref(profiling): Move profiling routes to utils
For clarity, move all the profiling route related functions to the utils folder.
